+++ ChangeLog.md
@@ -1,4 +1,4 @@
-## 0.0.3
+## Unreleased
 
 * Fall back to cabal dependency solver when a snapshot can't be found
 * Basic implementation of `stack new` [#137](https://github.com/commercialhaskell/stack/issues/137)